+++ benches/filters.rs
@@ -3,8 +3,22 @@ extern crate criterion;
 extern crate pdatastructs;
 
 use criterion::Criterion;
-use pdatastructs::cuckoofilter::CuckooFilter;
+use pdatastructs::filters::bloomfilter::BloomFilter;
+use pdatastructs::filters::cuckoofilter::CuckooFilter;
 use pdatastructs::rand::{ChaChaRng, SeedableRng};
+
+fn bloomfilter_add_single(c: &mut Criterion) {
+    c.bench_function("bloomfilter_add_single", |b| {
+        let false_positive_rate = 0.02; // = 2%
+        let expected_elements = 1000;
+        let mut filter = BloomFilter::with_properties(expected_elements, false_positive_rate);
+        let obj = "foo bar";
+
+        b.iter(|| {
+            filter.add(&obj);
+        })
+    });
+}
 
 fn cuckoofilter_add_many(c: &mut Criterion) {
     c.bench_function("cuckoofilter_add_many", |b| {
@@ -22,5 +36,5 @@ fn cuckoofilter_add_many(c: &mut Criterion) {
     });
 }
 
-criterion_group!(benches, cuckoofilter_add_many);
+criterion_group!(benches, bloomfilter_add_single, cuckoofilter_add_many);
 criterion_main!(benches);
