+++ src/filters/mod.rs
@@ -1,0 +1,4 @@
+//! Filters, Approximate Membership Queries (AMQs).
+
+pub mod bloomfilter;
+pub mod cuckoofilter;
